refactor(base): Deleted obsolete Themes
BREAKING CHANGE Deleted Themes `sap_belize`, `sap_belize_plus` and `sap_belize_hcb`
BREAKING CHANGE Removed `createThemeMap` from base
BREAKING CHANGE Removed `getThemeClassFor` from base
BREAKING CHANGE Removed `getThemeOverwrite` from base
BREAKING CHANGE Deleted `Themes` from base
