+++ packages/base/src/HSLColor.test.ts
@@ -1,30 +1,19 @@
 import { HSLColor } from './HSLColor';
-import { sapUiBaseText, sapAccentColor7, sapColorSemanticNegative, sapUiHighlight } from './sap_belize';
+import { sapColorSemanticNegative, sapUiBrand, sapUiHighlight } from './sap_fiori_3';
 
 describe('HSL Color', () => {
   test('lighten', () => {
-    const color = HSLColor.of(sapUiBaseText).lighten(20).hsl;
-    expect(color).toEqual('hsl(0, 0%, 40%)');
+    const color = HSLColor.of(sapUiHighlight).lighten(63).hsl;
+    expect(color).toEqual('hsl(209.8, 90.9%, 95%)'); //hsl(208.4, 90.5%, 95.9%)
   });
 
   test('darken', () => {
-    const color = HSLColor.of(sapAccentColor7).darken(8).hsl;
-    expect(color).toEqual('hsl(180, 70.8%, 26.9%)');
-  });
-
-  test('border colors', () => {
-    const color = HSLColor.of(sapColorSemanticNegative).lighten(10).hsl;
-    expect(color).toEqual('hsl(0, 100%, 46.7%)');
+    const color = HSLColor.of(sapUiBrand).darken(10).hsl;
+    expect(color).toEqual('hsl(209.8, 90.9%, 32.9%)'); //hsl(210, 90.5%, 32.9%)
   });
 
   test('border bg colors', () => {
-    const color = HSLColor.of(sapColorSemanticNegative).lighten(58).hsl;
-    expect(color).toEqual('hsl(0, 100%, 94.7%)');
-  });
-
-  test('Combined Saturate and Lighten', () => {
-    const saturated = HSLColor.of(sapUiHighlight).saturate(10).hsl;
-    const color = HSLColor.of(saturated).lighten(10).hsl;
-    expect(color).toEqual('hsl(207.2, 54.5%, 56%)'); // hsl(207.3, 54.8%, 56.7%)
+    const color = HSLColor.of(sapColorSemanticNegative).lighten(59.5).hsl;
+    expect(color).toEqual('hsl(0, 100%, 95.5%)'); //hsl(0, 100%, 96.1%)
   });
 });
