--- packages/base/src/Themes.ts
@@ -1,7 +1,0 @@
-enum Themes {
-  sap_belize = 'sap_belize',
-  sap_belize_plus = 'sap_belize_plus',
-  sap_belize_hcb = 'sap_belize_hcb'
-}
-
-export default Themes;
