+++ packages/base/src/Util.ts
@@ -1,39 +1,3 @@
-import Themes from './Themes';
-
-type ThemeMap = { [theme in Themes]: object };
-
-export const getThemeClassFor = (theme: Themes, themes: ThemeMap) => {
-  if ({}.hasOwnProperty.call(themes, theme)) {
-    return themes[theme];
-  }
-  throw new Error('Could not match theme to Theme Class!');
-};
-
-/**
- * Create a Theme Map for given styles
- * @param light {object} Belize Theme CSS Module
- * @param dark {object} Belize Deep Theme CSS Module
- * @param hcb {object} High Contrast Black Theme CSS Module
- * @return {ThemeMap} Theme Map
- */
-export const createThemeMap = (light: object, dark: object, hcb: object = {}): ThemeMap => ({
-  [Themes.sap_belize]: light,
-  [Themes.sap_belize_plus]: dark,
-  [Themes.sap_belize_hcb]: hcb
-});
-
-/**
- * Get overwrite theme class
- * @param {string} theme
- * @param {string} overwrite
- * @return {string} Theme String
- *
- */
-export const getThemeOverwrite = (theme: Themes, overwrite: Themes): Themes => {
-  if ((Object as any).values(Themes).includes(overwrite)) return overwrite;
-  return theme;
-};
-
 export const deprecationNotice = (component: string, message: string) => {
   const value = `* ui5-webcomponents-react Deprecation Notice - ${component}`;
   const dots = '*'.padStart(value.length, '*');
