+++ packages/base/src/index.ts
@@ -1,43 +1,29 @@
 /**
  * Created by d059190 at 16.03.18
  */
+import font72 from './72';
+import { bootstrap } from './bootstrap';
+import * as fonts from './fonts';
+import { HSLColor } from './HSLColor';
+import * as sap_fiori_3 from './sap_fiori_3';
+import * as spacing from './spacing';
+import { withStyles } from './withStyles';
+import { Device } from './Device';
 import { Event } from './Event';
+import { isNumeric } from './isNumeric';
 import KeyCodes from './KeyCodes';
 import { LOG_LEVEL, Logger } from './Logger';
 import Optional from './Optional';
 import StyleClassHelper from './StyleClassHelper';
-import Themes from './Themes';
-import { isNumeric } from './isNumeric';
-import { Device } from './Device';
-import * as fonts from '../../base/src/fonts';
-import font72 from '../../base/src/72';
-import * as sap_belize from '../../base/src/sap_belize';
-import * as sap_belize_plus from '../../base/src/sap_belize_plus';
-import * as sap_belize_hcb from '../../base/src/sap_belize_hcb';
-import * as sap_fiori_3 from '../../base/src/sap_fiori_3';
-import * as spacing from '../../base/src/spacing';
-import { HSLColor } from '../../base/src/HSLColor';
-import { bootstrap } from '../../base/src/bootstrap';
-import { withStyles } from '../../base/src/withStyles';
 
-import {
-  createThemeMap,
-  deprecationNotice,
-  getThemeClassFor,
-  getThemeOverwrite,
-  pushElementBackInScreen
-} from './Util';
+import { deprecationNotice, pushElementBackInScreen } from './Util';
 
 export {
   StyleClassHelper,
   KeyCodes,
   Optional,
-  getThemeClassFor,
-  createThemeMap,
-  getThemeOverwrite,
   deprecationNotice,
   pushElementBackInScreen,
-  Themes,
   Event,
   Logger,
   LOG_LEVEL,
@@ -45,9 +31,6 @@ export {
   Device,
   fonts,
   font72,
-  sap_belize,
-  sap_belize_plus,
-  sap_belize_hcb,
   spacing,
   HSLColor,
   sap_fiori_3,
