+++ packages/fiori3/src/components/ThemeProvider/index.tsx
@@ -1,12 +1,10 @@
-import { bootstrap, sap_fiori_3 } from '@ui5-webcomponents-react/base';
+import { sap_fiori_3 } from '@ui5-webcomponents-react/base';
 import { getCompactSize, getTheme } from '@ui5/webcomponents-base/src/Configuration';
 import React, { Fragment, PureComponent, ReactNode } from 'react';
 import { jss, ThemeProvider as ReactJssThemeProvider } from 'react-jss';
 import { ContentDensity } from '../../lib/ContentDensity';
 import { MessageToast } from '../../lib/MessageToast';
 import { Themes } from '../../lib/Themes';
-
-bootstrap();
 
 export interface ThemeProviderProps {
   withToastContainer?: boolean;
